% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asca_tds.R
\name{asca_tds}
\alias{asca_tds}
\title{Compute ANOVA Simultaneous Component Analysis (ASCA) of Temporal Dominant Sensation (TDS) data.}
\usage{
asca_tds(formula, data, timecol, attributes, time.quantization = NULL, ...)
}
\arguments{
\item{formula}{An object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted. The formula will be applied for each atttribut at each time interval defined by the timecol column}

\item{data}{A data frame, or object coercible by as.data.frame to a data frame, containing the variables in the model. It must be in long format and must contain a column for Temporal Dominant Sensation binary data, a column reporting the time values, and a column that defines the attributes analyzed.}

\item{timecol}{A string containing the name of the column indicating the time intervals of the TDS dataset.}

\item{attributes}{A string containing the name of the column indicating the attributes of the TDS.}

\item{time.quantization}{A single number reporting the number of time units contained in the new intervals used to quantize the time column.}

\item{...}{Optional parameters}
}
\value{
A list of objects containing the results of ASCA decomposition of a structured dataset.
}
\description{
Compute ANOVA Simultaneous Component Analysis (ASCA) of Temporal Dominant Sensation (TDS) data.
}
\examples{
\dontrun{
asca_tds(CATA~(sample+assessor)^2, data = tempR::bars \%>\%
gather( time, CATA, 5:455),timecol = "time", attributes = "attribute")
describe(dataset, col1, col2)
}
}
