% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asca_ti.R
\name{asca_ti}
\alias{asca_ti}
\title{Compute ANOVA Simultaneous Component Analysis (ASCA) of Time Intensity (TI) data.}
\usage{
asca_ti(formula, data, timecol, time.quantization = NULL, ...)
}
\arguments{
\item{formula}{An object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted. The formula will be applied for each atttribute at each time interval defined by the timecol column}

\item{data}{A data frame, or object coercible by as.data.frame to a data frame, containing the variables in the model. It must be in long format and must contain a column for Temporal Dominant Sensation binary data, a column reporting the time values, and a column that defines the attributes analyzed.}

\item{timecol}{A string containing the name of the column indicating the time intervals of the TI dataset.}

\item{time.quantization}{A single number reporting the number of time units contained in the new intervals used to quantize the time column.}

\item{...}{Optional parameters}
}
\value{
A list of objects containing the results of ASCA decomposition of a structured dataset.
}
\description{
Compute ANOVA Simultaneous Component Analysis (ASCA) of Time Intensity (TI) data.
}
\examples{
\dontrun{

#first of all a dataset of Time Intensity data is used.
# An open Time Intensity dataset is downloaded from the website
# https://help.xlstat.com/dataset/time-intensity-data_0.xlsm
library(readxl)
library(tidyr)
library(stringr)
library(httr)
url1 <- "https://help.xlstat.com/dataset/time-intensity-data_0.xlsm"
GET(url1, write_disk(tf <- tempfile(fileext = ".xlsm")))
tf <- str_replace_all(tf, "\\\\\\\\", "//")
data <- read_excel(tf)
data.long <- data \%>\%
pivot_longer(names_to = "time", values_to = "intensity", cols = 6:36) \%>\%
droplevels()

asca_ti(intensity~PRODUCT*PANELIST,
data = data.long,timecol = "time")
}
}
