% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_asca.R
\name{perm_asca}
\alias{perm_asca}
\title{Compute permutation test to validate the results from ANOVA Simultaneous Component Analysis (ASCA) of Temporal Check All That Apply (TCATA) data.}
\usage{
perm_asca(
  data,
  ASCA_object,
  nrep = 1000,
  plot = F,
  confidence = 0.95,
  test = c("permutation", "bootstrap"),
  ...
)
}
\arguments{
\item{data}{A data frame, or object coercible by as.data.frame to a data frame, containing the variables in the model. It must be in long format and must contain a column for Temporal Check All That Apply binary data, a column reporting the time values, and a column that defines the attributes analyzed.}

\item{ASCA_object}{An object obtained from the functions asca_ti(), asca_tds(), and asca_tcata() from the package ASCATCATA. The model selected must be th one estimated from the data value.}

\item{nrep}{Number of iteration for bootstrap and permutation test, default is 1000.}

\item{plot}{Logical. If TRUE, prints a plot after estimating the permutation test.}

\item{confidence}{Numeric, values from 0 to 1. Indicates the probability applied for the estimation of the confidence interval via bootstrapping.}

\item{test}{A vector of strings defyning which kind of iteration test will be performed by the function between permutation test (to estimate the statistical significance of the model) and bootstrap test (to define the confidence intervals of the parameters estimated). The application of permutation depends on the presence of the string "permutation" in the vector, and the application of bootstrap dpends on the presence of the string "bootstrap".}

\item{...}{Optional parameters}
}
\value{
A list of objects containing the results of ASCA decomposition of a structured dataset.
}
\description{
Compute permutation test to validate the results from ANOVA Simultaneous Component Analysis (ASCA) of Temporal Check All That Apply (TCATA) data.
}
\examples{
\dontrun{
perm_asca(CATA~(samp+cons)^2, data = tempR::ojtcata  \%>\%
gather(time, CATA, 5:25) \%>\%
 mutate(cons = as.factor(cons), samp = as.factor(samp), type = "tcata",
timecol = "time", attributes = "attribute", nrep = 1000)
}

}
