% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_tcatasca.R
\name{perm_tcatasca}
\alias{perm_tcatasca}
\title{Compute permutation test to validate the results from ANOVA Simultaneous Component Analysis (ASCA) of Temporal Check All That Apply (TCATA) data.}
\usage{
perm_tcatasca(data, formula, timecol, attributes, ...)
}
\arguments{
\item{data}{A data frame, or object coercible by as.data.frame to a data frame, containing the variables in the model. It must be in long format and must contain a column for Temporal Check All That Apply binary data, a column reporting the time values, and a column that defines the attributes analyzed.}

\item{formula}{An object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted. The formula will be applied for each atttribut at each time interval defined by the timecol column}

\item{timecol}{A string containing the name of the column indicating the time intervals of the TCATA dataset.}

\item{attributes}{A string containing the name of the column indicating the attributes of the TCATA.}

\item{...}{Optional parameters}
}
\value{
A list of objects containing the results of ASCA decomposition of a structured dataset.
}
\description{
Compute permutation test to validate the results from ANOVA Simultaneous Component Analysis (ASCA) of Temporal Check All That Apply (TCATA) data.
}
\examples{
\dontrun{
perm_tcatasca(CATA~(samp+cons)^2, data = tempR::ojtcata, time, attribute)
describe(dataset, col1, col2)
}

}
