% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bootstrap.R
\name{plot_bootstrap}
\alias{plot_bootstrap}
\title{Creates bi-plots similar to those created by the \code{plot_ASCA()} function, but it includes also the confidence intervals estimated by the bootstrapped values estimated by the \code{perm_asca()} function.}
\usage{
plot_bootstrap(
  ASCA_obj,
  bootstrap_obj,
  object = NA,
  time.label = NULL,
  print = T,
  score.points = T,
  score.labels = T,
  dimensions = c(1, 2),
  path = T,
  density = F,
  path.smooth = T,
  h_clus = NULL,
  max.overlaps.value = 10,
  point.size = 2
)
}
\arguments{
\item{ASCA_obj}{A list of PCA estimated by the TCATASCA function.}

\item{bootstrap_obj}{A list of estimated values from bootstrap test estimated using the perm_asca() function.}

\item{object}{A vector of string or numbers indicating which parameter of the ASCA decomposition will be plotted}

\item{time.label}{Numeric. Adds numeric values of the time units in the plot for loadings values for loadings.time.structure == "long". The number indicates how big has to be the interval between each label.}

\item{print}{Logical. Indicates whether or not to print the plots.}

\item{score.points}{Logical. Adds the points indicating the scores values. Standard is TRUE.}

\item{score.labels}{Logical. Adds the text labels indicating the score names. Standard is TRUE.}

\item{dimensions}{A numeric vector indicating two numbers indicating the axes of the ASCA decomposition.}

\item{path}{Logical. Adds a path indicating the position of the loadings according to their cronological order. Default is TRUE.}

\item{density}{Logical. Superimpose a 2-dimensional density plot, indicating th distribution of the temporal resolved loadings according to their density. Default is FLASE.}

\item{path.smooth}{Logical. Adds a path smoothed indicating the position of the loadings according to their cronological order. Default is TRUE.}

\item{h_clus}{Numeric. Indicates whether to calculate or not hierchical clustering for the levels of each factor. The algorithm applied for hierarchical clustering is "Ward-D2", and it is estimated from euclidean distance for all the principal component estimated.}

\item{max.overlaps.value}{Numeric, default is 10. Define the maximum number of overlaps allowed for text in the plots.}

\item{point.size}{A numeric value defining the size of the points of the score values.}
}
\value{
A series of plots representing the results of the permutation test and the results of the bootstrap tests.
}
\description{
Creates bi-plots similar to those created by the \code{plot_ASCA()} function, but it includes also the confidence intervals estimated by the bootstrapped values estimated by the \code{perm_asca()} function.
}
\examples{

\dontrun{
#'ASCA_model <- asca_tcata(CATA~(samp+cons)^2, data = data.long, timecol = "time",
        attributes = "attribute")

perm_model <- perm_asca(data = data.long, ASCA_object = ASCA_model,
nrep = 1000, test = "bootstrap")

plot_bootstrap(ASCA_obj = ASCA_model, bootstrap_obj = perm_model)

}




}
