% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_permutation.R
\name{plot_permutation}
\alias{plot_permutation}
\title{Creates}
\usage{
plot_permutation(
  ASCA_obj,
  perm_object,
  axes = c(1, 2),
  point.size = 2,
  max.overlaps.value = 10,
  print = T
)
}
\arguments{
\item{ASCA_obj}{A list of PCA estimated by the TCATASCA function.}

\item{perm_object}{A list of estimated values from permutation test and bootstrap test estimated using the perm_asca() function.}

\item{axes}{A numeric vector indicating two numbers indicating the axes of the ASCA decomposition.}

\item{point.size}{A numeric value defining the size of the points of the score values.}

\item{max.overlaps.value}{Numeric, default is 10. Define the maximum number of overlaps allowed for text in the plots.}

\item{print}{Logical. Indicates whether or not to print the plots.}
}
\value{
A series of plots representing the results of the permutation test and the results of the bootstrap tests.
}
\description{
Creates
}
\examples{

\dontrun{
#'ASCA_model <- asca_tcata(CATA~(samp+cons)^2, data = data.long, timecol = "time",
        attributes = "attribute")
perm_model <- perm_asca(data = data.long, ASCA_object = ASCA_model,
nrep = 1000)

plot_permutation(perm_model)

}




}
