% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_time_loadings.R
\name{plot_time_loadings}
\alias{plot_time_loadings}
\title{Compute ANOVA Simultaneous Component Analysis (ASCA) of Temporal Check All That Apply (TCATA) data.}
\usage{
plot_time_loadings(
  ASCA_obj,
  choice = "contrib",
  ref = c("attributes", "factors"),
  object = NA,
  print = TRUE,
  lab_two_lvl = FALSE,
  dimensions = c(1, 2)
)
}
\arguments{
\item{ASCA_obj}{A list of PCA estimated by the TCATASCA function.}

\item{choice}{A string that could be "loadings" or "contrib", to indicate whether the loading values or the contribution values will be estimatd and plotted.}

\item{ref}{A string vector indicating which graph print. "attributes" prints a plot using the attributes factor as a faceting variable. "factor" prints a plot with the attribute variable as a faceting factor.}

\item{object}{A vector of string or numbers indicating which parameter of the ASCA decomposition will be plotted}

\item{print}{Logical. Indicates wether or not to print the plots.}

\item{lab_two_lvl}{Logical. Adds a label indicating the direction where the score values are if in the plot there are loading values from a factor with only two levels. Default is FALSE.}

\item{dimensions}{A numeric vector indicating two numbers indicating the dimensions of the ASCA decomposition.}
}
\value{
A plot representing the loadings of the ASCA decomposition and the density plot of the loadings of ASCA decomposition.
}
\description{
Compute ANOVA Simultaneous Component Analysis (ASCA) of Temporal Check All That Apply (TCATA) data.
}
\examples{
\dontrun{
# The function creates two plots from an ASCA_obj in long format

ASCA_obj <- asca_tcata(CATA~(samp+cons)^2, data = tempR::ojtcata  \%>\%
gather(time, CATA, 5:25) \%>\%
 mutate(cons = as.factor(cons), samp = as.factor(samp),
        time = as.numeric(str_extract(time, "\\\\d+"))), timecol = "time",
        attributes = "attribute")

plot_time_loadings(ASCA_obj)

# it is possible to define which are the factor considered using the object
# parameter, to define wich factors will be reported in the plots

plot_time_loadings(ASCA_obj, object = 1:2)

#On default settings, the function reports the estimation of the contribution
# on the overall variance of each attribute among time.
# To report the loading values it is necessary to define `choice` = "loadings",
# once it is defined, the function reports a line plot presenting the loading values
# among time for each attribute.

plot_time_loadings(ASCA_obj, choice = "loadings")

# It is possible to define for which dimensions the contribution or
# the loading values will be reported. You can define them at the dimensions variable

plot_time_loadings(ASCA_obj, dimensions = c(2,3))

}
}
