% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_tcatasca.R
\name{SS_tcatasca}
\alias{SS_tcatasca}
\title{Estimate the cumulative sum of squares for each factor included in an ANOVA Simultaneous Component Analysis (ASCA) of Temporal Check All That Apply (TCATA) data, to estimate the effect size of the decomposition.}
\usage{
SS_tcatasca(data, formula, timecol, attributes, graph = T, ...)
}
\arguments{
\item{data}{A data frame, or object coercible by as.data.frame to a data frame, containing the variables in the model. It must be in long format and must contain a column for Temporal Check All That Apply binary data, a column reporting the time values, and a column that defines the attributes analyzed.}

\item{formula}{An object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted. The formula will be applied for each atttribut at each time interval defined by the timecol column}

\item{timecol}{A string containing the name of the column indicating the time intervals of the TCATA dataset.}

\item{attributes}{A string containing the name of the column indicating the attributes of the TCATA.}

\item{graph}{Logical. If TRUE prints a barplot showing the percentage of cumulative sum of squares related to each factor.}

\item{...}{Optional parameters}
}
\value{
A list of objects containing the results of ASCA decomposition of a structured dataset.
}
\description{
Estimate the cumulative sum of squares for each factor included in an ANOVA Simultaneous Component Analysis (ASCA) of Temporal Check All That Apply (TCATA) data, to estimate the effect size of the decomposition.
}
\examples{
\dontrun{
SS_tcatasca(CATA~(samp+cons)^2, data = tempR::ojtcata, time, attribute)
describe(dataset, col1, col2)
}

}
