% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ASCA.R
\name{plot_ASCA}
\alias{plot_ASCA}
\title{Creates bi-plots for each factor decompsed in the ASCA object elaborated by the function tcatasca().}
\usage{
plot_ASCA(
  ASCA_obj,
  object = NA,
  print = T,
  axes = c(1, 2),
  path = T,
  density = F,
  path.smooth = T,
  h_clus = NULL,
  point.size = 2
)
}
\arguments{
\item{ASCA_obj}{A list of PCA estimated by the TCATASCA function.}

\item{object}{A vector of string or numbers indicating which parameter of the ASCA decomposition will be plotted}

\item{print}{Logical. Indicates wether or not to print the plots.}

\item{axes}{A numeric vector indicating two numbers indicating the axes of the ASCA decomposition.}

\item{path}{Logical. Adds a path indicating the position of the loadings according to their cronological order. Default is TRUE.}

\item{density}{Logical. Superimpose a 2-dimensional density plot, indicating th distribution of the temporal resolved loadings according to their density. Default is FLASE.}

\item{path.smooth}{Logical. Adds a path smoothed indicating the position of the loadings according to their cronological order. Default is TRUE.}

\item{h_clus}{Numeric. Indicates wether to calculate or not hierchical clustering for the levels of each factor. The algorithm applied for hierarchical clustering is "Ward-D2", and it is estimated from euclidean distance for all the principal component estimated.}

\item{point.size}{A numeric value defining the size of the points of the score values.}
}
\value{
A series of plots representing the scores of the ASCA decomposition and the values of the loadings of the same ASCA decomposition.
}
\description{
Creates bi-plots for each factor decompsed in the ASCA object elaborated by the function tcatasca().
}
\examples{
\dontrun{
plot_ASCA(ASCA_obj, object = 1:3)
}
}
