% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ASCA.R
\name{plot_ASCA}
\alias{plot_ASCA}
\title{Creates bi-plots for each factor decompsed in the ASCA object elaborated by the function tcatasca().}
\usage{
plot_ASCA(
  ASCA_obj,
  object = NA,
  print = T,
  axes = c(1, 2),
  path = T,
  density = F,
  path.smooth = T,
  h_clus = NULL,
  max.overlaps.value = 10,
  point.size = 2
)
}
\arguments{
\item{ASCA_obj}{A list of PCA estimated by the TCATASCA function.}

\item{object}{A vector of string or numbers indicating which parameter of the ASCA decomposition will be plotted}

\item{print}{Logical. Indicates wether or not to print the plots.}

\item{axes}{A numeric vector indicating two numbers indicating the axes of the ASCA decomposition.}

\item{path}{Logical. Adds a path indicating the position of the loadings according to their cronological order. Default is TRUE.}

\item{density}{Logical. Superimpose a 2-dimensional density plot, indicating th distribution of the temporal resolved loadings according to their density. Default is FLASE.}

\item{path.smooth}{Logical. Adds a path smoothed indicating the position of the loadings according to their cronological order. Default is TRUE.}

\item{h_clus}{Numeric. Indicates wether to calculate or not hierchical clustering for the levels of each factor. The algorithm applied for hierarchical clustering is "Ward-D2", and it is estimated from euclidean distance for all the principal component estimated.}

\item{max.overlaps.value}{Numeric, default is 10. Define the maximum number of overlaps allowed for text in the plots.}

\item{point.size}{A numeric value defining the size of the points of the score values.}
}
\value{
A series of plots representing the scores of the ASCA decomposition and the values of the loadings of the same ASCA decomposition.
}
\description{
Creates bi-plots for each factor decompsed in the ASCA object elaborated by the function tcatasca().
}
\examples{
\dontrun{
library(tidyverse)
# According to the type and the structure of the ASCA_obj inserted in the function
# there will be displayed an according plot.

# For ASCA_TCATA and ASCA_TDS in long structure the data will be
# displayed using bi-plots for
# each factor included in the ASCA decomposition.

ASCA_obj <- asca_tcata(CATA~(samp+cons)^2, data = tempR::ojtcata  \%>\%
gather(time, CATA, 5:25) \%>\%
 mutate(cons = as.factor(cons), samp = as.factor(samp),
        time = as.numeric(str_extract(time, "\\\\d+"))), timecol = "time",
        attributes = "attribute")

plot_ASCA(ASCA_obj, object = 1:3)

# For ASCA_TCATA and ASCA_TDS in short structure, the results
# will be displayed in line plots for each component, with barplots
# reporting the loading values

ASCA_obj <- asca_tcata(CATA~(samp+cons)^2, data = tempR::ojtcata  \%>\%
gather(time, CATA, 5:25) \%>\%
 mutate(cons = as.factor(cons), samp = as.factor(samp),
        time = as.numeric(str_extract(time, "\\\\d+"))), timecol = "time",
        attributes = "attribute", loadings.time.structure = "short")

plot_ASCA(ASCA_obj)

# It is possible to define only one or more factors to be reported using
# `object`, can be defined using a string or a numerical value.

plot_ASCA(ASCA_obj, object = "cons")

# For ASCA_TI object the results will be reported in biplots

ASCA_obj <- asca_ti(CATA~(sample+assessor)^2,
data = read_excel("time-intensity-data_0.xlsm") \%>\%
gather(time, intensity, 6:36), timecol = "time")

plot_ASCA(ASCA_obj)

# using the `axes` parameter is possible to specify which axes will be
# displayed in the plot

plot_ASCA(ASCA_obj, axes = c(2,3))

# Estethic parameters can be addd or removed specifying TRUE or FALSE towards
 # the parameters density, path, and path.smooth

plot_ASCA(ASCA_obj, density = TRUE, path = FALSE, path.smooth = FALSE)

# for data in long structure it is possible to apply a hierachical
clustering procedure indicating the number of desidered clusters in h_clus+

plot_ASCA(ASCA_obj, h_clus = 3)

}
}
