% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_time_loadings.R
\name{plot_time_loadings}
\alias{plot_time_loadings}
\title{Compute ANOVA Simultaneous Component Analysis (ASCA) of Temporal Check All That Apply (TCATA) data.}
\usage{
plot_time_loadings(
  ASCA_obj,
  ref = c("attributes", "factors"),
  object = 1:(length(ASCA_obj) - 2),
  print = T,
  axes = c(1, 2)
)
}
\arguments{
\item{ASCA_obj}{A list of PCA estimated by the TCATASCA function.}

\item{ref}{A string vector indicating which graph print. "attributes" prints a plot using the attributes factor as a faceting variable. "factor" prints a plot with the attribute variable as a faceting factor.}

\item{object}{A vector of string or numbers indicating which parameter of the ASCA decomposition will be plotted}

\item{print}{Logical. Indicates wether or not to print the plots.}

\item{axes}{A numeric vector indicating two numbers indicating the axes of the ASCA decomposition.}
}
\value{
A plot representing the loadings of the ASCA decomposition and the density plot of the loadings of ASCA decomposition.
}
\description{
Compute ANOVA Simultaneous Component Analysis (ASCA) of Temporal Check All That Apply (TCATA) data.
}
\examples{
\dontrun{
plot_time_loadings(ASCA_obj, object = 1:3, axes = c(1,2), print = T)
}
}
